Use /health/service/ instead of /catalog/service/
It turns out that there is no way to only get services that are passing
their health check from the /catalog endpoints, the way to do it is to
use /health. This goes back to the fact that the Consul catalog isn't
really designed for ephemeral nodes, so it assumes that you'd want all
addresses, even if they are not healthy.

Because consul uses different JSON formats for the same resource in
different endpoints, I had to change some of the JSON parsing code.

See: https://www.consul.io/docs/agent/http/health.html#health_service